requirements.txt
anyio==3.6.2
asyncio==3.4.3
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
click==8.1.3
colorama==0.4.6
ecdsa==0.18.0
fastapi==0.92.0
h11==0.14.0
idna==3.4
psycopg2==2.9.5
pyasn1==0.4.8
pydantic==1.10.5
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
wincertstore==0.2
